pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kubeagentic"
version = "0.1.0"
description = "A Python library for building LangGraph agents from YAML configuration with OpenAI-compatible REST API"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "KubeAgentic Contributors"}
]
keywords = ["langgraph", "langchain", "agent", "yaml", "openai", "api", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core LangGraph/LangChain
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-community>=0.3.0",
    
    # Configuration & Validation
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "jsonschema>=4.17",
    
    # API Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0.0",
    
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    
    # Database & Caching
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "redis>=5.0",
    "psycopg2-binary>=2.9",
    
    # Monitoring & Observability
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    
    # Utilities
    "tenacity>=8.2",
    "aiofiles>=24.0",
    "python-json-logger>=2.0",
    "httpx>=0.27.0",
    "tiktoken>=0.7.0",
    
    # CLI
    "typer>=0.12.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "httpx>=0.27.0",
    "black>=24.0",
    "ruff>=0.6.0",
    "mypy>=1.11",
    "pre-commit>=3.8",
]

test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "httpx>=0.27.0",
    "locust>=2.31",
]

docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
]

all = [
    "kubeagentic[dev,test,docs]",
]

[project.scripts]
kubeagentic = "kubeagentic.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/kubeagentic"
Documentation = "https://kubeagentic.readthedocs.io"
Repository = "https://github.com/yourusername/kubeagentic"
Issues = "https://github.com/yourusername/kubeagentic/issues"

[tool.setuptools]
packages = ["kubeagentic"]

[tool.setuptools.package-data]
kubeagentic = ["py.typed", "schemas/*.json", "examples/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=kubeagentic",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"

[[tool.mypy.overrides]]
module = [
    "langgraph.*",
    "langchain.*",
    "tiktoken.*",
]
ignore_missing_imports = true 